split methods in `binder/expr/function`
